objstore: initial mockup of object store shared library
The idea is to abstract all the details of volume and volume group
management in a shared library.  Then, both the client and server daemons
can use the same library to manage their respective local storage.

The object store will have two components: (1) a generic API that consumers
use, and (2) an assortment of backends responsible for persistently storing
objects.

Currently, there are no backends.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>