common: keep track of the local node
This commit defines two functions: one to set the local node ID and one to
retrieve it.  When the common lib gets loaded, the local node ID is initialized
to zero.  (Recall that this is not a valid node ID.)  The set function allows
one to set the node once.  If you wish to change the node id, you'll have to
restart the process.  This allows anyone linking against the common lib to
assume that the node id will not change on them.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>