objstore: simple module loading on init, and most of create operation
The objstore uses loadable modules to implement different backends.  The loading
and linking is done via dlopen(3).

The create operation itself is pretty simple.  It sets up the generic objstore
sturcture and calls the backend.  The backend needs to complete initialization
of the struct as well as do whatever other setup is necessary.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>