cmake: bins should be executable, libs may need to be writable
CMake's "make install" installs the file, sets the permissions, and then optionally
proceeds to remove the RPATH from the file.  This fails if the file is not
writable because of the permissions.  In the ideal world, cmake would remove
the RPATH first, but since we don't live in an ideal world, we'll just have to
install certain files as writable by the owner.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>