objstore/mem: link count is not a property of obj versions
The link count (nlink) is a property of an object regardless of how many
versions it may have.  (Recall that dentries point to objects not object
versions.)  Therefore we must keep track of links on per-object basis.
Since we're using struct nattr for "disk" layout even though it is not meant
for that, we cannot remove the nlink field from it.  So instead we set the
version's nattr nlink to an obviously bad value, and keep track of the real
link count using a separate variable in the object itself.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>