objstore/mem: implement open and close object operations
We use the memver struct pointer as the cookie returned from open.  In
close, we then cast the passed in cookie and dereference it.  This is safe,
because the consumers of the objstore library are trusted.  It is up to them
to wrap the cookie as needed before presenting it to untrusted clients
(e.g., over the network).

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>