objstore/mem: fill any holes with zeroes when extending a file via a write
GNU cp(1) tries to be clever and creates sparse files.  As a result, I've
seen the following syscall sequence:

	read(3, "7F E L F0101010601\0\0\0".., 131072)	= 12632
	write(4, "7F E L F0101010601\0\0\0".., 4608)	= 4608
	llseek(4, 3584, SEEK_CUR)			= 8192
	write(4, " T  06\b\0\0\0\0\0\0\0\0".., 4440)	= 4440

Note that the first write(2) writes at offset 0, the llseek(2) seeks to 8192
(4608+3584), and the second write(2) writes 4440 bytes at this 8k offset
resulting in a file that is 12632 bytes long.  The interesting thing here is
the 3584 byte "hole" that's implicitly zeroed.

As the mem backend only receives WRITE RPCs with offsets, it will extend the
file twice.  First to 4608 bytes and then to 12632 bytes.  The second time
around, it needs to zero the hole at [4608,8192).

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>