objstore/mem: use the vdev uuid generated by common code
Instead of using our own uuid, use the uuid generated by the common code.

Note that this is technically wrong, since we need a volume uuid here but
we're using a vdev uuid.  This will get fixed soon once the backends are
fixed to understand the difference between vdevs and vols.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>