common: don't use negative numbers for NERR_RPC_ERROR and NERR_UNKNOWN_ERROR
Using negative numbers seemed like a good idea, but if we keep all of them
positive we can use an int to return either a negated errno or a NERR
without any ambiguities.

The choice of numbers just shy of 64k is intentional - it allows error codes
to be stored in a u16.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>