common: split fscall_connect into two: connect and mount
Some code (e.g., the admin utility) will want to connect but don't have a
particular volume to "mount".  Splitting the code into two functions will
let the admin utility reuse the connection setup part.

Additionally, the connect part now only does a handshake with the server.
It is up to the consumer to establish a socket to pass in.  This is a tiny
bit more work for the consumer, but it also makes it painfully clear whether
we failed to establish a connection or there was a handshare issue.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>